watson: implement HTTP text2speech transcription
Fixes #1.

HTTP transcription in that it is a single request from
which we then download back the transcribed audio.

Details are given in the watson/README.md file, but
for the purposes of illustration here is an example:

```go
package main

import (
	"fmt"
	"io"
	"log"
	"os"
	"time"

	"github.com/orijtech/text2speech/watson"
)

func main() {
	client, err := watson.NewClientFromEnv()
	if err != nil {
		log.Fatal(err)
	}

	f, err := os.Create("output-audio.wav")
	if err != nil {
		log.Fatal(err)
	}
	defer f.Close()

	now := time.Now()
	hour, min, sec := now.Clock()
	text := fmt.Sprintf("The time in 24hr clock is %v:%v:%v", hour, min, sec)
	rc, err := client.SynthesizeAudio(&watson.Request{
		Text:  text,
		Voice: watson.VoiceFrenchFemale,

		OutputContentType: watson.WAV,
	})
	if err != nil {
		log.Fatal(err)
	}
	defer rc.Close()

	io.Copy(f, rc)
}
```

and that will save a file "output-audio.wav" in your working directory
from which you can then listen to the audio.